fixed issue on balancedTileSize

+++ tiling/experiments/papi/jacobi/ParallelTiled.c
@@ -62,7 +62,7 @@ int main(int argc, char const *argv[])
 
     //getting works performance here. Check
 	// initAndMeasure(&eventSet, event);
-	int balancedTileSize = (sqrt(cacheSize*0.7/4)/cores);
+	int balancedTileSize = (sqrt( (double) (cacheSize*0.7/4) )/cores);
 	if ((((1+(-2*m))+(m*m))<=100000)&&(cacheSize>((8*m)*m)))
 	{
 		#pragma loop name main #0
