+++ tiling/experiments/speed-up/matrix-multiplication/ParallelSingleTiled.c
@@ -62,7 +62,7 @@ int main(int argc, char const *argv[])
     int i, j, k;
     int _ret_val_0;
 
-	int balancedTileSize = (sqrt(cacheSize*0.7/4)/cores);
+	int balancedTileSize = (sqrt( (double) (cacheSize*0.7/4) )/cores);
     double start = omp_get_wtime();
 
 	if ((((m*n)*n)<=100000)&&(cacheSize>(((8*m)*n)+((4*n)*n))))
